# backend/src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: ちいぴよ API
  description: 子供の写真・動画管理及び育児記録管理アプリケーションのAPI定義
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /trash:
    get:
      tags:
        - trashManagement
      summary: ゴミ箱内のメディア一覧を取得
      operationId: getTrashItems
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: ゴミ箱メディア一覧
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/TrashItemListResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - trashManagement
      summary: ゴミ箱からメディアを完全に削除
      operationId: deleteTrashItems
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/TrashItemIds'
      responses:
        '204':
          description: メディアの完全削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - trashManagement
      summary: ゴミ箱からメディアを復元
      operationId: restoreTrashItems
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/TrashRestoreRequestDto'
      responses:
        '204':
          description: メディアの復元に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /trash/empty:
    delete:
      tags:
        - trashManagement
      summary: ゴミ箱の中身をすべて空にする
      operationId: emptyTrash
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      responses:
        '204':
          description: すべてのメディアの完全削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /trash/{id}:
    delete:
      tags:
        - trashManagement
      summary: ゴミ箱からメディアを完全に削除
      operationId: deleteTrashItem
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: 完全削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /trash/{id}/restore:
    post:
      tags:
        - trashManagement
      summary: ゴミ箱からメディアを復元
      operationId: restoreTrashItem
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: メディアの復元に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognitoで取得したJWTトークンを指定
  parameters:
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        default: XMLHttpRequest
    ResourceId:
      name: id
      in: path
      description: リソースの一意な識別子
      required: true
      schema:
        type: integer
        format: int64
    PageOffset:
      name: offset
      in: query
      description: 取得開始位置
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    PageLimit:
      name: limit
      in: query
      description: 取得件数
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    TrashItemIds:
      name: trashItemIds
      in: query
      description: ゴミ箱データのIDリスト
      required: true
      schema:
        type: array
        items:
          type: integer
          format: int64
  schemas:
    ErrorResponseDto:
      type: object
      description: エラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストが不正です
...
    TagResponseDto:
      type: object
      description: タグ情報
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: タグID
          example: 1
        name:
          type: string
          description: タグ名
          example: お出かけ
        mediaCount:
          type: integer
          format: int64
          description: タグが紐づいているメディアの数
          example: 10
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
    MediaNavigationResponseDto:
      type: object
      description: 前後のメディア情報
      required:
        - id
        - mediaType
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3アップロード用の署名付きURL
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
    MediaResponseDto:
      type: object
      description: メディア情報
      required:
        - id
        - uploadedBy
        - mediaType
        - originalFilename
        - contentType
        - fileSize
        - width
        - height
        - takenAt
        - albumId
        - sharingGroupId
        - uploadStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        uploadedBy:
          type: integer
          format: int64
          description: アップロードしたユーザーID
          example: 1
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
        originalFilename:
          type: string
          description: ファイル名
          example: IMG_20260101.jpg
        contentType:
          type: string
          description: MIMEタイプ
          example: image/jpeg
        fileSize:
          type: integer
          format: int64
          description: ファイルサイズ（バイト）
          example: 2048000
        width:
          type:
            - integer
            - 'null'
          description: 幅（ピクセル）
          example: 1920
        height:
          type:
            - integer
            - 'null'
          description: 高さ（ピクセル）
          example: 1080
        presignedUrl:
          type: string
          format: uri
          description: S3署名付きURL
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3署名付きURL
        takenAt:
          type:
            - string
            - 'null'
          format: date
          description: 撮影日
          example: '2026-01-01'
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
          example: 1
        sharingGroupId:
          type: integer
          format: int64
          description: 共有範囲グループID
          example: 1
        uploadStatus:
          type: string
          description: アップロード状況
          enum:
            - PROCESSING
            - COMPLETED
            - FAILED
          example: COMPLETED
        addFavoriteUserIds:
          type: array
          description: お気に入りに追加したユーザーID一覧
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
        tags:
          type: array
          description: 付与されているタグ一覧
          items:
            $ref: '#/components/schemas/TagResponseDto'
        isFavorite:
          type: boolean
          description: お気に入り済みフラグ
          example: false
        commentCount:
          type: integer
          format: int64
          description: コメント数
          example: 1
        nextMedia:
          type:
            - object
            - 'null'
          description: 1つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondNextMedia:
          type:
            - object
            - 'null'
          description: 2つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        previousMedia:
          type:
            - object
            - 'null'
          description: 1つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondPreviousMedia:
          type:
            - object
            - 'null'
          description: 2つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
...
    TrashItemResponseDto:
      type: object
      description: ゴミ箱メディア情報
      required:
        - id
        - media
        - expiresAt
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: ゴミ箱ID
          example: 1
        media:
          $ref: '#/components/schemas/MediaResponseDto'
        expiresAt:
          type: string
          format: date-time
          description: 完全削除予定日時
          example: '2026-02-01T00:00:00Z'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
    TrashItemListResponseDto:
      type: object
      description: ゴミ箱一覧レスポンス
      required:
        - items
        - earliest
        - totalCount
        - hasNext
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrashItemResponseDto'
        earliest:
          type:
            - integer
            - null
          format: int64
          description: 最も期限が近いメディアの完全削除までの残り日数
          example: 3
        totalCount:
          type: integer
          format: int64
          description: 総件数
          example: 5
        hasNext:
          type: boolean
          description: 次のページが存在するかどうか
          example: true
    TrashRestoreRequestDto:
      type: object
      description: 複数ゴミ箱復元リクエスト
      required:
        - trashItemIds
      properties:
        trashItemIds:
          type: array
          items:
            type: integer
            format: int64
          description: ゴミ箱メディアのID配列
          example:
            - 1
            - 2
            - 3
  responses:
    DefaultErrors:
      description: 予期しないエラーが発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Unauthorized:
      description: 認証エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    NotFound:
      description: 指定されたリソースが見つからない
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    InternalServerError:
      description: サーバー内部でエラー発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
...
    Forbidden:
      description: アクセス権限エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'